#143 Remove HistoryLogService interface
